Image feedback (#608)
* Allow the use of generated images as inputs


# Conflicts:
#	google/generativeai/vision_models/_vision_models.py

* types + formatting


* add files


* Fix 3.9


* Fix 3.9


* fix pytype


* fix pytype


* typo


* reapply commits lost in merge


* Update google/generativeai/client.py

* Remove GCS reference


* black .
